move functions and components into their own files
