+++ examples/simple/src/tags/TagEdit.tsx
@@ -32,7 +32,7 @@ const TagEdit = props => (
                 basePath="/posts"
                 resource="posts"
                 filter={{ tags: [props.id] }}
-                title={<span />}
+                title={false}
             >
                 <Datagrid>
                     <TextField source="id" />
