+++ packages/ra-core/src/util/index.ts
@@ -10,7 +10,8 @@ import Ready from './Ready';
 import resolveRedirectTo from './resolveRedirectTo';
 import warning from './warning';
 import useWhyDidYouUpdate from './useWhyDidYouUpdate';
-import { useSafeSetState, useTimeout, useDocumentTitle } from './hooks';
+import useDocumentTitle from './useDocumentTitle';
+import { useSafeSetState, useTimeout } from './hooks';
 export * from './indexById';
 
 export {
