+++ packages/ra-ui-materialui/src/layout/Title.tsx
@@ -2,24 +2,14 @@ import * as React from 'react';
 import { FC, cloneElement, ReactElement } from 'react';
 import { createPortal } from 'react-dom';
 import PropTypes from 'prop-types';
-import { useTranslate, Record, warning, useDocumentTitle } from 'ra-core';
+import { useTranslate, Record, warning } from 'ra-core';
+import TitleText from './TitleText';
 
 export interface TitleProps {
     className?: string;
     defaultTitle?: string;
     record?: Record;
-    title?: string | ReactElement;
-}
-
-interface TitleTextProps {
-    text: string;
-    className: string;
-}
-
-function TitleText({ text, ...rest }: TitleTextProps) {
-    useDocumentTitle(text);
-
-    return <span {...rest}>{text}</span>;
+    title?: string | ReactElement | false;
 }
 
 const Title: FC<TitleProps> = ({
@@ -36,6 +26,9 @@ const Title: FC<TitleProps> = ({
             : null;
 
     if (!container) return null;
+
+    if (typeof title === 'boolean' && !title) return null;
+
     warning(!defaultTitle && !title, 'Missing title prop in <Title> element');
 
     const titleElement = !title ? (
@@ -55,12 +48,14 @@ const Title: FC<TitleProps> = ({
 export const TitlePropType = PropTypes.oneOfType([
     PropTypes.string,
     PropTypes.element,
+    PropTypes.bool,
 ]);
 
 Title.propTypes = {
     defaultTitle: PropTypes.string,
     className: PropTypes.string,
     record: PropTypes.any,
+    // @ts-ignore
     title: TitlePropType,
 };
